[Chore] Separate out `vllm.utils.func`
